grid cells can now have different drained area, implemented export to gtiff using rasterio

+++ PyOCN/c_src/ocn.c
@@ -37,7 +37,7 @@ Status ocn_update_energy(FlowGrid *G, drainedarea_t da_inc, linidx_t a, double g
     }
     // update root vertex
     da = vert.drained_area;
-    G->energy += pow((double)(da + da_inc), gamma) - pow((double)da, gamma);  // update energy
+    G->energy += pow((da + da_inc), gamma) - pow(da, gamma);  // update energy
     vert.drained_area += da_inc;  // update drained area of vertex
     fg_set_lin(G, vert, a);
